example-subgraph: Move generate types to the correct dir
